made changes to the secrets.json had a typo in the environment

+++ AlejandroElectronics/Controllers/ShippingsController.cs
@@ -7,15 +7,21 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using AlejandroElectronics.Models;
 
+
+
 namespace AlejandroElectronics.Controllers
 {
     public class ShippingsController : Controller
     {
         private readonly AlejandroTestContext _context;
-        
-        public ShippingsController(AlejandroTestContext context)
+       
+        private Braintree.BraintreeGateway _braintreeGateway; // => injecting braintree
+
+        public ShippingsController(AlejandroTestContext context, Braintree.BraintreeGateway braintreeGateway)
         {
             _context = context;
+
+            _braintreeGateway = braintreeGateway;
         }
         
         
@@ -49,7 +55,8 @@ namespace AlejandroElectronics.Controllers
 
         
         [HttpPost]
-        public async Task<IActionResult> Index(ShippingsViewModel model) // this Model info needs to go to braintree.
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Index(ShippingsViewModel model, string CreditCardNumber, string CreditCardName, string ExpirationYear, string ExpirationMonth, string CreditCardVerificationValue) // this Model info needs to go to braintree.
         {
             
             if (Request.Cookies.Keys.Contains("cartId") && Guid.TryParse(Request.Cookies["cartId"], out Guid cartId))
@@ -74,11 +81,41 @@ namespace AlejandroElectronics.Controllers
                         State = "IL"
                     }
                 });
-                _context.Orders.Add(newOrder); // => this adds the newly created order.
-                _context.Cart.Remove(model.Cart); // => once it's added, Remove() will clear out the cart.
-                Response.Cookies.Delete("cartId");// => Delete() will delete the cookie with the cart info.
-                await _context.SaveChangesAsync();
-                return RedirectToAction("Index", "OrderComplete");
+                if (ModelState.IsValid)
+                {
+
+
+                    Braintree.TransactionRequest saleRequest = new Braintree.TransactionRequest();
+                    saleRequest.Amount = model.Cart.Product.Price.Value;
+                    saleRequest.CreditCard = new Braintree.TransactionCreditCardRequest
+                    {
+                        CardholderName = model.CreditCartName,
+                        CVV = model.CreditCardVerificationValue,
+                        ExpirationMonth = model.ExpirationMonth,
+                        ExpirationYear = model.ExpirationYear,
+                        Number = model.CreditCardNumber,
+
+                    };
+
+                    // awaiting the result of card validation
+                    var result = await _braintreeGateway.Transaction.SaleAsync(saleRequest);
+                    if (result.IsSuccess())
+                    {
+                        _context.Orders.Add(newOrder); // => this adds the newly created order.
+                        _context.Cart.Remove(model.Cart); // => once it's added, Remove() will clear out the cart.
+                        Response.Cookies.Delete("cartId");// => Delete() will delete the cookie with the cart info.
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction("Index", "OrderComplete");
+
+                    }
+                    foreach (var error in result.Errors.All())
+                    {
+                        ModelState.AddModelError(error.Code.ToString(), error.Message);
+                    }
+                }
+
+               
+                //return RedirectToAction("Index", "OrderComplete");
             }
